update install function

+++ zsh/installer.sh
@@ -1,11 +1,5 @@
 #!/bin/sh
 
-JudgeZshrcExist(){
-    if [ -d "~/.zshrc" ]; then
-        mv ~/.zshrc ~/.zshrc_back
-    fi
-}
-
 InstallZsh(){
     echo "installing..."
     if which apt-get >/dev/null; then
@@ -15,29 +9,38 @@ InstallZsh(){
     elif which yum >/dev/null;then
         sudo yum install zsh git python3 neovim npm -y
     elif which pacman >/dev/null;then
-        sudo pacman -S --noconfirm zsh git python3 neovim xclip xorg-xclipboard npm trash-cli
+        sudo pacman -Syy && sudo pacman -S --noconfirm zsh git python3 neovim xclip xorg-xclipboard npm trash-cli
     else
         echo "Does not support system version"
     fi
 }
 
 InstallOhMyZsh(){
-    git clone https://github.com/robbyrussell/oh-my-zsh.git ~/.oh-my-zsh
-    ln -s ~/.dotfiles/zsh/.zshrc ~/.zshrc
-    chsh -s $(which zsh)
+    if [ -e ~/.oh-my-zsh ]; then
+        echo "~/.oh-my-zsh exists"
+    else
+        git clone $1/robbyrussell/oh-my-zsh.git ~/.oh-my-zsh
+    fi
+
+    if [ -e ~/.zshrc ];then
+        echo "~/.zshrc exists"
+    else
+        ln -s ~/.dotfiles/zsh/.zshrc ~/.zshrc
+        chsh -s $(which zsh)
+    fi
 }
 
 InstallThemesPlugins(){
-    git clone https://github.com/zsh-users/zsh-autosuggestions ~/.oh-my-zsh/plugins/zsh-autosuggestions
-    git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ~/.oh-my-zsh/plugins/zsh-syntax-highlighting
+	echo $1
+    git clone $1/zsh-users/zsh-autosuggestions ~/.oh-my-zsh/plugins/zsh-autosuggestions
+    git clone $1/zsh-users/zsh-syntax-highlighting.git ~/.oh-my-zsh/plugins/zsh-syntax-highlighting
     ln -s ~/.dotfiles/zsh/Schminitz.zsh-theme ~/.oh-my-zsh/custom/themes/Schminitz.zsh-theme
 }
 
 
-JudgeZshrcExist
 InstallZsh
-InstallOhMyZsh
-InstallThemesPlugins
+InstallOhMyZsh $1
+InstallThemesPlugins $1
 
 echo "edit ~/.oh-my-zsh/themes/Schminitz.zsh-themes to update themes"
 echo "finish ! logout and relogin"
